Heatmap of the genes and groups

Generate a heatmap with classification worthy genes and cancer groups

hm_gpa_sel(expDat, genes, grps, maxPerGrp = 100, cRow = FALSE,
  cCol = FALSE, limits = c(0, 10), toScale = FALSE,
  fontsize_row = 4, reOrderCells = FALSE, ...)

Arguments

expDat
normalized expression matrix from trans_prop

genes
classification worth genes from findClassyGenes

grps
a vector that maps sample names to a cancer category

maxPerGrp
the number max samples per group

cRow
TRUE if the rows should be clustered

cCol
TRUE if the columns should be clustered

limits
a vector of size 2 indicating the lowest and highest range for expression level

toScale
TRUE if the data should be scaled

fontsize_row
the font size for the row labels

Value

a heatmap of genes and their groups